requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.33
botocore==1.38.33
certifi==2025.1.31
cfn-flip==1.3.0
charset-normalizer==3.4.1
click==8.2.1
distro==1.9.0
durationpy==0.10
exceptiongroup==1.2.2
Flask==3.1.1
h11==0.16.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
kappa==0.6.0
lxml==5.3.2
MarkupSafe==3.0.2
openai==1.86.0
outcome==1.3.0.post0
packaging==25.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
placebo==0.9.0
pydantic==2.11.5
pydantic_core==2.33.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.13.0
selenium==4.31.0
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
troposphere==4.9.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
zappa==0.60.2
